requirements.txt
crewai
langchain_google_genai
google-generativeai
load_dotenv
crewai_tools